Create test suite

+++ Test/typing/should_fail/same_args_in_method.java
@@ -1,7 +1,7 @@
 package tutu.titi.toto;
 
 class A {
-	public int method1(int i, float i) {
+	public int method1(int i, int i) {
 		return 1;
 	}
 }
